Fixed the core-2815 and core-2816

+++ api/src/main/java/org/ednovo/gooru/domain/service/CourseRepositoryHibernate.java
@@ -1,4 +1,26 @@
-
+/////////////////////////////////////////////////////////////
+// CourseRepositoryHibernate.java
+// gooru-api
+// Created by Gooru on 2015
+// Copyright (c) 2015 Gooru. All rights reserved.
+// http://www.goorulearning.org/
+// Permission is hereby granted, free of charge, to any person obtaining
+// a copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to
+// permit persons to whom the Software is furnished to do so, subject to
+// the following conditions:
+// The above copyright notice and this permission notice shall be
+// included in all copies or substantial portions of the Software.
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+/////////////////////////////////////////////////////////////
 package org.ednovo.gooru.domain.service;
 
 import java.util.List;
@@ -14,17 +36,17 @@ import org.springframework.stereotype.Repository;
 public class CourseRepositoryHibernate extends BaseRepositoryHibernate implements CourseRepository, ParameterProperties, ConstantProperties {
 
 	@Override
-	public Course getCourse(Short courseId) {
+	public Course getCourse(Integer courseId) {
 		Query query = getSession().createQuery("FROM Course c  WHERE c.courseId=:courseId").setParameter("courseId", courseId);
 		return (Course) (query.list().size() > 0 ? query.list().get(0) : null);
 	}
-	
+
 	@Override
 	public Course getCourseCode(String courseCode) {
 		Query query = getSession().createQuery("FROM Course c  WHERE c.courseCode=:courseCode").setParameter("courseCode", courseCode);
 		return (Course) (query.list().size() > 0 ? query.list().get(0) : null);
 	}
-	
+
 	@Override
 	public List<Course> getCourses(Integer limit, Integer offset) {
 		Query query = getSession().createQuery("FROM Course");
@@ -36,8 +58,7 @@ public class CourseRepositoryHibernate extends BaseRepositoryHibernate implement
 	@Override
 	public Long getCourseCount() {
 		Query query = getSession().createQuery("SELECT COUNT(*) FROM Course");
-		return (Long) (query.list().size() > 0 ? query.list().get(0) : 0);		
+		return (Long) (query.list().size() > 0 ? query.list().get(0) : 0);
 	}
 
-	
 }
